added auth

+++ app/Http/Controllers/AdminModelController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\AdminModel;
+use App\Models\Auth;
 use App\Http\Controllers\URL;
 use Session;
 
@@ -13,16 +14,19 @@ use Session;
 class AdminModelController extends Controller
 {
     function index(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       return view('adminindex');
     }
 
     function games(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $games = AdminModel::getGames();
       // dump($games);
       return view('admingames',['games'=>$games]);
     }
 
     function gameDetails($id){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $game = AdminModel::getGameDetails($id);
       $genreList = AdminModel::getGenres();
       // $genreListArray = array('');
@@ -44,6 +48,7 @@ class AdminModelController extends Controller
     }
 
     function submitGameChanges(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       // dd(request()->post());
       $response = request()->post();
       $gameId = $response['IdGame'];
@@ -57,11 +62,13 @@ class AdminModelController extends Controller
     }
 
     function pricing(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $pricing = AdminModel::getPricing();
       return view('adminpricing', ['pricing' => $pricing]);
     }
 
     function submitPricingChanges(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $response = request()->post();
       $gamePrice = $response['gamePrice'];
       $consolePrice = $response['consolePrice'];
@@ -69,20 +76,24 @@ class AdminModelController extends Controller
     }
 
     function orders(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       echo "orders";
     }
 
     function console(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $consoles = AdminModel::getConsoles();
       return view('adminconsoles',['consoles'=>$consoles]);
     }
 
     function consoleDetails($id){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $console = AdminModel::getConsoleDetails($id);
       return view('adminconsoledetails', ['console' => $console]);
     }
 
     function submitConsoleChanges(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $response = request()->post();
       $idConsole = $response['ConsoleID'];
       $namaConsole = $response['NamaConsole'];
@@ -93,12 +104,14 @@ class AdminModelController extends Controller
     }
 
     function addnewgame(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $genreList = AdminModel::getGenres();
       $platformList = AdminModel::getPlatformList();
       return view('adminaddnewgame', ['genreList' => $genreList, 'platformList' => $platformList]);
     }
 
     function submitnewgame(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $response = request()->post();
       $namaGame = $response['namaGame'];
       $platform = $response['platform'];
@@ -109,10 +122,12 @@ class AdminModelController extends Controller
     }
 
     function addnewconsole(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       return view('adminaddnewconsole');
     }
 
     function submitnewconsole(){
+      if(Auth::authAdmin()) return redirect()->route('home');
       $response = request()->post();
       $namaConsole = $response['NamaConsole'];
       $qty = $response['qty'];
@@ -121,10 +136,12 @@ class AdminModelController extends Controller
     }
 
     function deletegame($id){
+      if(Auth::authAdmin()) return redirect()->route('home');
       return AdminModel::deleteGame($id);
     }
 
     function deleteconsole($id){
+      if(Auth::authAdmin()) return redirect()->route('home');
       return AdminModel::deleteConsole($id);
     }
